[Add] Implement useSortedAndFilteredUsers hook for improved user sorting and filtering [Style] Add filter-container class for better input styling in UserTable
